solved merge issues
