--- app/src/main/java/com/ux/ok_baby/Constants.java
@@ -1,15 +1,0 @@
-package com.ux.ok_baby;
-
-public class Constants {
-    static final String FOOD = "Food";
-    static final String SLEEP = "Sleep";
-    static final String DIAPER = "Diaper";
-    static final String OTHER = "Other";
-    static final int FOOD_TAB = 1;
-    static final int SLEEP_TAB = 0;
-    static final int DIAPER_TAB = 2;
-    static final int OTHER_TAB = 3;
-    static final String REPORT_TYPE = "report type";
-    static final String BABY_ID = "baby id";
-    static final String DATE_PATTERN = "MM/dd/yy";
-}
